✨ Add hidden way to fetch all episodes
If the requested page is `-1`, all the episodes will be fetched.
This is not ideal as it'll perform many requests, but it may be useful
if one needs access to very old episodes, without iterating over all
the pages.

NB: "hidden" in the sense not displayed on the html website.