🐛 Fix undefined in description
The bodyMarkdown field was recently added to the API, without
the standfirst one being filled. This caused the description to be
undefined.

Some descriptions are not at all available, in no fields (and from the
radio france app, it's also missing). For those, instead of undefined,
put an empty description.