🐛 Ignore missing shows from relationship
With some input (`science`), the parsing of the shows was failing
as some shows were missing from the relationship object.

Address it by ignoring them. As it's the first report like that,
I believe we can safely ignore them, but it may happen in the
future that we ignored a true show. To be tackled at that future
time.

Fix #8